// src/app/classes/app-state.interface.ts
export interface ListingViewState {
  order: Array<string | number>;
  page?: LookupTable<Array<number | string>>;
  total?: number;
  query?: Query;
  loading: LookupTable<boolean>;
}
// ...
export interface LookupTable<T> {
  [id: string]: T;
}